+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Cacher/Interface/ICache.cs
@@ -1,7 +1,5 @@
-﻿namespace OxGFrame.AssetLoader
+﻿namespace OxGFrame.AssetLoader.Cacher
 {
-    public delegate void Progression(float progress, float reqSize, float totalSize);
-
     public interface ICache<T>
     {
         bool HasInCache(string name);