+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/GroupCacher/GroupBundle.cs
@@ -64,7 +64,6 @@ namespace OxGFrame.AssetLoader.GroupCahcer
                 if (keyGroup != null)
                 {
                     keyGroup.AddRef();
-
                     Logging.Print<Logger>($"【Load】 => Current << GroupBundle >> Cache Count: {this.Count}, KeyRef: {keyGroup.refCount}, GroupId: {id}");
                 }
             }
@@ -85,7 +84,6 @@ namespace OxGFrame.AssetLoader.GroupCahcer
                 if (keyGroup != null)
                 {
                     keyGroup.AddRef();
-
                     Logging.Print<Logger>($"【Load】 => Current << GroupBundle >> Cache Count: {this.Count}, KeyRef: {keyGroup.refCount}, GroupId: {id}");
                 }
             }
@@ -99,21 +97,18 @@ namespace OxGFrame.AssetLoader.GroupCahcer
             if (keyGroup != null)
             {
                 keyGroup.DelRef();
-
                 Logging.Print<Logger>($"【Unload】 => Current << GroupBundle >> Cache Count: {this.Count}, KeyRef: {keyGroup.refCount}, GroupId: {id}");
 
                 // 強制釋放
                 if (forceUnload)
                 {
                     this.DelFromCache(id, keyGroup.assetName);
-
                     Logging.Print<Logger>($"【Force Unload Completes】 => Current << GroupBundle >> Cache Count: {this.Count}, GroupId: {id}");
                 }
                 // 使用引用計數釋放
                 else if (keyGroup.refCount <= 0)
                 {
                     this.DelFromCache(id, keyGroup.assetName);
-
                     Logging.Print<Logger>($"【Unload Completes】 => Current << GroupBundle >> Cache Count: {this.Count}, GroupId: {id}");
                 }
 
@@ -192,7 +187,6 @@ namespace OxGFrame.AssetLoader.GroupCahcer
                 if (keyGroup != null)
                 {
                     keyGroup.AddRef();
-
                     Logging.Print<Logger>($"【Load】 => Current << GroupBundle >> Cache Count: {this.Count}, KeyRef: {keyGroup.refCount}, GroupId: {id}");
                 }
             }
@@ -213,7 +207,6 @@ namespace OxGFrame.AssetLoader.GroupCahcer
                 if (keyGroup != null)
                 {
                     keyGroup.AddRef();
-
                     Logging.Print<Logger>($"【Load】 => Current << GroupBundle >> Cache Count: {this.Count}, KeyRef: {keyGroup.refCount}, GroupId: {id}");
                 }
             }
@@ -227,21 +220,18 @@ namespace OxGFrame.AssetLoader.GroupCahcer
             if (keyGroup != null)
             {
                 keyGroup.DelRef();
-
                 Logging.Print<Logger>($"【Unload】 => Current << GroupBundle >> Cache Count: {this.Count}, KeyRef: {keyGroup.refCount}, GroupId: {id}");
 
                 // 強制釋放
                 if (forceUnload)
                 {
                     this.DelFromCache(id, keyGroup.assetName);
-
                     Logging.Print<Logger>($"【Force Unload Completes】 => Current << GroupBundle >> Cache Count: {this.Count}, GroupId: {id}");
                 }
                 // 使用引用計數釋放
                 else if (keyGroup.refCount <= 0)
                 {
                     this.DelFromCache(id, keyGroup.assetName);
-
                     Logging.Print<Logger>($"【Unload Completes】 => Current << GroupBundle >> Cache Count: {this.Count}, GroupId: {id}");
                 }