+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Progression.cs
@@ -1,0 +1,4 @@
+﻿namespace OxGFrame.AssetLoader
+{
+    public delegate void Progression(float progress, float currentCount, float totalCount);
+}